add config file, routes.py
