+++ config.py.tmpl
@@ -1,0 +1,7 @@
+from os.path import dirname, join
+
+
+class Config(object):
+    SECRET_KEY = 'super top secret key'
+    UPLOAD_FOLDER='input'
+    OUTPUT_FOLDER='output'
